Use double rather than int to store priorities
This is a better match for the common Prioritized interface and Priority
utility class.

+++ core/tools/src/main/java/imagej/core/tools/FloodFillTool.java
@@ -61,7 +61,7 @@ public class FloodFillTool extends AbstractTool {
 
 	// -- instance variables --
 
-	public static final int PRIORITY = -304;
+	public static final double PRIORITY = -304;
 
 	enum Connectivity {
 		EIGHT, FOUR
